+++ pyvo/dal/exceptions.py
@@ -50,7 +50,7 @@ class DALAccessError(Exception):
         return self._reason
 
     def __repr__(self):
-        return "{}: {}".format(self._typeName(self), self._reason)
+        return f"{self._typeName(self)}: {self._reason}"
 
     @property
     def reason(self):
@@ -84,9 +84,7 @@ class DALProtocolError(DALAccessError):
         ----------
         reason : str
            a message describing the cause of the error
-        code : int
-           the HTTP error code (as an integer)
-        cause : str
+        cause : Exception
            an exception issued as the underlying cause.  A value
            of None indicates that no underlying exception was
            caught.
@@ -117,7 +115,7 @@ class DALFormatError(DALProtocolError):
 
         Parameters
         ----------
-        cause : str
+        cause : Exception
            an exception issued as the underlying cause.  A value
            of None indicates that no underlying exception was caught.
         url
@@ -150,7 +148,7 @@ class DALServiceError(DALProtocolError):
            a message describing the cause of the error
         code : int
            the HTTP error code (as an integer)
-        cause : str
+        cause : Exception
            an exception issued as the underlying cause.  A value
            of None indicates that no underlying exception was
            caught.
@@ -188,13 +186,13 @@ class DALServiceError(DALProtocolError):
                 code = response.status_code
                 content_type = response.headers.get('content-type', None)
                 if content_type and 'text/plain' in content_type:
-                    message = '{} for {}'.format(response.text, url)
+                    message = f'{response.text} for {url}'
 
             # TODO votable handling
 
             return DALServiceError(message, code, exc, url)
         elif isinstance(exc, Exception):
-            return DALServiceError("{}: {}".format(cls._typeName(exc), str(exc)),
+            return DALServiceError(f"{cls._typeName(exc)}: {str(exc)}",
                                    cause=exc, url=url)
         else:
             raise TypeError("from_except: expected Exception")