+++ CHANGES.rst
@@ -4,12 +4,35 @@
 Enhancements and Fixes
 ----------------------
 
+- Extending the MIVOT module with the ability to build annotations component by component
+  and put them into a VOTable. [#627]
+
+- Make deletion of TAP jobs optional via a new ``delete`` kwarg. [#640]
+
+- Change AsyncTAPJob.result to return None if no result is found explicitly [#644]
+
+- Add a UAT constraint to the registry interface for constraining
+  subjects [#649]
+
 
 Deprecations and Removals
 -------------------------
 
+- Versions of Python <3.9 are no longer supported. [#639]
 
-1.6.1 (unreleased)
+
+1.6.2 (2025-04-07)
+==================
+
+Bug Fixes
+---------
+
+- Fix performance issues with datalink results. [#654]
+
+- More careful NULL value handling in tapregext data limits. [#659]
+
+
+1.6.1 (2025-02-12)
 ==================
 
 Bug Fixes
@@ -18,6 +41,10 @@ Bug Fixes
 - Fix propagating some previously swallowed exceptions. [#614]
 
 - Fix string literal generation for SQL query when using numpy >=2.0. [#624]
+
+- Switch to do minimal VOSI tables downloads for TAP metadata. [#634]
+
+- Provide more informative exception message when requests to endpoints fail. [#641]
 
 
 1.6 (2024-11-01)
@@ -51,23 +78,18 @@ Enhancements and Fixes
 - MIVOT module: the model references in the dictionaries that are used to build ``MivotInstance``
   objects are made more consistent [#551]
 
-- MIVOT module: If the MIVOT annotation block contains a valid instance of the
-  ``mango:EpochPosition`` class, the dynamic object describing the mapped
-  data can generate a valid SkyCoord instance. [#591]
-
-- Extending the MIVOT module with the ability to build annotations component by component
-  and put them into a VOTable. [#627]
-  
 - RegTAP constraints involving tables other than rr.resource are now
   done via subqueries for less duplication of interfaces. [#562, #572]
 
+- MIVOT module: If the MIVOT annotation block contains a valid instance of the
+  ``mango:EpochPosition`` class, the dynamic object describing the mapped
+  data can generate a valid SkyCoord instance. [#591]
 
 - New sub-package discover for global dataset discovery. [#470]
 
 - Updated getdatalink to be consistent with iter_datalinks. [#613]
 
-  
-  
+
 Deprecations and Removals
 -------------------------